+++ src/Machina/Drivers/Communication/Protocols/ABBCommunicationProtocol.cs
@@ -31,11 +31,12 @@ namespace Machina.Drivers.Communication.Protocols
         internal const int RES_POSE = 21;                       // ">21 400 300 500 0 0 1 0;"
         internal const int RES_JOINTS = 22;                     // ">22 0 0 0 0 90 0;"
         internal const int RES_EXTAX = 23;                      // ">23 1000 9E9 9E9 9E9 9E9 9E9;" Sends external axes values
+        internal const int RES_FULL_POSE = 24;                  // ">24 X Y Z QW QX QY QZ J1 J2 J3 J4 J5 J6 A1 A2 A3 A4 A5 A6;" Sends all pose and joint info (probably on split messages)
 
-        
+
         // Characters used for buffer parsing
         internal const char STR_MESSAGE_END_CHAR = ';';       // Marks the end of a message
-        internal const char STR_MESSAGE_CONTINUE_CHAR = '>';  // Marks the end of an unfinished message, to be continued on next message
+        internal const char STR_MESSAGE_CONTINUE_CHAR = '>';  // Marks the end of an unfinished message, to be continued on next message. Useful when the message is too long and needs to be split in chunks
         internal const char STR_MESSAGE_ID_CHAR = '@';        // Flags a message as an acknowledgment message corresponding to a source id
         internal const char STR_MESSAGE_RESPONSE_CHAR = '$';  // Flags a message as a response to an information request(acknowledgments do not include it)