+++ src/Machina/Drivers/Communication/TCPCommunicationManagerABB.cs
@@ -17,23 +17,27 @@ namespace Machina.Drivers.Communication
     /// </summary>
     internal class TCPCommunicationManagerABB
     {
-        internal TCPConnectionStatus Status { get; private set; }
+        
 
-        private RobotCursor _writeCursor;
+        private RobotCursor _releaseCursor;
+        private RobotCursor _executionCursor;
         private RobotCursor _motionCursor;
+
         private Driver _parentDriver;
         internal RobotLogger logger;
 
-        private const int INIT_TIMEOUT = 10000;  // in millis
+        private const int INIT_TIMEOUT = 5000;  // in millis
         internal Vector initPos;
         internal Rotation initRot;
         internal Joints initAx;
         internal ExternalAxes initExtAx;
 
-        private TcpClient clientSocket = new TcpClient();
-        private NetworkStream clientNetworkStream;
-        private Thread receivingThread;
-        private Thread sendingThread;
+        // Properties for Driver module
+        private TcpClient _clientSocket;
+        private NetworkStream _clientNetworkStream;
+        private TCPConnectionStatus _clientStatus;
+        private Thread _receivingThread;
+        private Thread _sendingThread;
         private string _ip = "";
         public string IP => _ip;
         private int _port = 0;
@@ -53,19 +57,35 @@ namespace Machina.Drivers.Communication
         private int _maxStreamCount = 10;
         private int _sendNewBatchOn = 2;
 
-        private bool _bufferEmptyEventIsRaiseable = true;
-
         private string _deviceDriverVersion = null;
 
+        // Props for Monitor module (if available)
+        private TcpClient _monitorClientSocket;
+        private TCPConnectionStatus _monitorStatus;
+        private Thread _monitoringThread;
+        private string _monitorIP;
+        private int _monitorPort;
+        private int _monitorReceiveByteCount;
+        private string _monitorMessage;
+        private byte[] _monitorReceiveMsgBytes = new byte[1024];
 
-        internal TCPCommunicationManagerABB(Driver driver, RobotCursor writeCursor, RobotCursor motionCursor, string ip, int port)
+        private bool _isMonitored = false;
+        public bool IsMonitored => _isMonitored;
+
+
+
+
+        internal TCPCommunicationManagerABB(Driver driver, RobotCursor releaseCursor, RobotCursor executionCursor, string ip, int port)
         {
             this._parentDriver = driver;
             this.logger = driver.parentControl.logger;
-            this._writeCursor = writeCursor;
-            this._motionCursor = motionCursor;
+            this._releaseCursor = releaseCursor;
+            this._executionCursor = executionCursor;
+            this._motionCursor = driver.parentControl.MotionCursor;   // @TODO: homogeinize how the driver picks cursors from parent control: as constructor arguments or directly from the object 
             this._ip = ip;
             this._port = port;
+            this._monitorIP = ip;               
+            this._monitorPort = port + 1;  // these should be configurable...   
 
             this._translator = Protocols.Factory.GetTranslator(this._parentDriver);
         }
@@ -73,15 +93,15 @@ namespace Machina.Drivers.Communication
         internal bool Disconnect()
         {
 
-            if (clientSocket != null)
+            if (_clientSocket != null)
             {
-                Status = TCPConnectionStatus.Disconnected;
-                clientSocket.Client.Disconnect(false);
-                clientSocket.Close();
-                if (clientNetworkStream != null) clientNetworkStream.Dispose();
+                _clientStatus = TCPConnectionStatus.Disconnected;
+                _clientSocket.Client.Disconnect(false);
+                _clientSocket.Close();
+                if (_clientNetworkStream != null) _clientNetworkStream.Dispose();
                 return true;
             }
-
+            
             return false;
         }
 
@@ -89,20 +109,20 @@ namespace Machina.Drivers.Communication
         {
             try
             {
-                clientSocket = new TcpClient();
-                clientSocket.Connect(this._ip, this._port);
-                Status = TCPConnectionStatus.Connected;
-                clientNetworkStream = clientSocket.GetStream();
-                clientSocket.ReceiveBufferSize = 1024;
-                clientSocket.SendBufferSize = 1024;
+                _clientSocket = new TcpClient();
+                _clientSocket.Connect(this._ip, this._port);
+                _clientStatus = TCPConnectionStatus.Connected;
+                _clientNetworkStream = _clientSocket.GetStream();
+                _clientSocket.ReceiveBufferSize = 1024;
+                _clientSocket.SendBufferSize = 1024;
 
-                sendingThread = new Thread(SendingMethod);
-                sendingThread.IsBackground = true;
-                sendingThread.Start();
+                _sendingThread = new Thread(SendingMethod);
+                _sendingThread.IsBackground = true;
+                _sendingThread.Start();
 
-                receivingThread = new Thread(ReceivingMethod);
-                receivingThread.IsBackground = true;
-                receivingThread.Start();
+                _receivingThread = new Thread(ReceivingMethod);
+                _receivingThread.IsBackground = true;
+                _receivingThread.Start();
 
                 if (!WaitForInitialization())
                 {
@@ -110,7 +130,9 @@ namespace Machina.Drivers.Communication
                     return false;
                 }
 
-                return clientSocket.Connected;
+                TryConnectMonitor();
+
+                return _clientSocket.Connected;
             }
             catch (Exception ex)
             {
@@ -120,6 +142,55 @@ namespace Machina.Drivers.Communication
 
             //return false;
         }
+
+        private bool TryConnectMonitor()
+        {
+            try
+            {
+                _monitorClientSocket = new TcpClient();
+                _monitorClientSocket.Connect(this._monitorIP, this._monitorPort);
+                _monitorStatus = TCPConnectionStatus.Connected;
+                _monitorClientSocket.ReceiveBufferSize = 1024;
+                _monitorClientSocket.SendBufferSize = 1024;
+
+                _monitoringThread = new Thread(MonitoringMethod);
+                _monitoringThread.IsBackground = true;
+                _monitoringThread.Start();
+
+                _isMonitored = true;
+
+                return _monitorClientSocket.Connected;
+            }
+            catch (Exception ex)
+            {
+                logger.Info("Real-time monitoring not available on this device");
+                DisconnectMonitor();
+            }
+
+            return false;
+        }
+
+
+        private bool DisconnectMonitor()
+        {
+            if (_monitorClientSocket != null)
+            {
+                if (_monitorClientSocket.Connected)
+                {
+                    _monitorClientSocket.Client.Disconnect(false);
+                }
+                _monitorClientSocket.Close();
+                _monitorStatus = TCPConnectionStatus.Disconnected;
+                _isMonitored = false;
+
+                return true;
+            }
+
+            return false;
+        }
+
+
+
 
         private bool WaitForInitialization()
         {
@@ -149,21 +220,21 @@ namespace Machina.Drivers.Communication
         {
             if (Thread.CurrentThread.Name == null)
             {
-                Thread.CurrentThread.Name = "MachinaTCPSendingThread";
+                Thread.CurrentThread.Name = "MachinaTCPDriverSendingThread";
             }
 
             // Expire the thread on disconnection
-            while (Status != TCPConnectionStatus.Disconnected)
+            while (_clientStatus != TCPConnectionStatus.Disconnected)
             {
-                while (this.ShouldSend() && this._writeCursor.AreActionsPending())
+                while (this.ShouldSend() && this._releaseCursor.AreActionsPending())
                 {
-                    var msgs = this._translator.GetMessagesForNextAction(this._writeCursor);
+                    var msgs = this._translator.GetMessagesForNextAction(this._releaseCursor);
                     if (msgs != null)
                     {
                         foreach (var msg in msgs)
                         {
                             _sendMsgBytes = Encoding.ASCII.GetBytes(msg);
-                            clientNetworkStream.Write(_sendMsgBytes, 0, _sendMsgBytes.Length);
+                            _clientNetworkStream.Write(_sendMsgBytes, 0, _sendMsgBytes.Length);
                             _sentMessages++;
                             //Console.WriteLine("Sending mgs: " + msg);
                             logger.Debug($"Sent: {msg}");
@@ -184,15 +255,15 @@ namespace Machina.Drivers.Communication
         {
             if (Thread.CurrentThread.Name == null)
             {
-                Thread.CurrentThread.Name = "MachinaTCPListeningThread";
+                Thread.CurrentThread.Name = "MachinaTCPDriverListeningThread";
             }
 
             // Expire the thread on disconnection
-            while (Status != TCPConnectionStatus.Disconnected)
+            while (_clientStatus != TCPConnectionStatus.Disconnected)
             {
-                if (clientSocket.Available > 0)
+                if (_clientSocket.Available > 0)
                 {
-                    _receiveByteCount = clientSocket.GetStream().Read(_receiveMsgBytes, 0, _receiveMsgBytes.Length);
+                    _receiveByteCount = _clientSocket.GetStream().Read(_receiveMsgBytes, 0, _receiveMsgBytes.Length);
                     _response = Encoding.UTF8.GetString(_receiveMsgBytes, 0, _receiveByteCount);
 
                     //var msgs = _response.Split(new char[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
@@ -209,6 +280,34 @@ namespace Machina.Drivers.Communication
                 Thread.Sleep(30);
             }
         }
+
+
+        private void MonitoringMethod(object obj)
+        {
+            if (Thread.CurrentThread.Name == null)
+            {
+                Thread.CurrentThread.Name = "MachinaTCPMonitorListeningThread";
+            }
+
+            while (_monitorStatus != TCPConnectionStatus.Disconnected)
+            {
+                if (_monitorClientSocket.Available > 0)
+                {
+                    _monitorReceiveByteCount = _monitorClientSocket.GetStream().Read(_monitorReceiveMsgBytes, 0, _monitorReceiveMsgBytes.Length);
+                    _monitorMessage = Encoding.UTF8.GetString(_monitorReceiveMsgBytes, 0, _monitorReceiveByteCount);
+
+                    var msgs = SplitResponse(_monitorMessage);
+                    foreach(var msg in msgs)
+                    {
+                        logger.Debug("Received message from monitor: " + msg);
+                        // @TODO: do something with the message: update MotionCursor
+                    }
+                }
+
+                Thread.Sleep(30);
+            }
+        }
+
 
         private bool ShouldSend()
         {
@@ -322,7 +421,7 @@ namespace Machina.Drivers.Communication
             _responseChunks = res.Split(' ');
             string idStr = _responseChunks[0].Substring(1);
             int id = Convert.ToInt32(idStr);
-            this._motionCursor.ApplyActionsUntilId(id);
+            this._executionCursor.ApplyActionsUntilId(id);
 
             // Raise appropriate events
             this._parentDriver.parentControl.RaiseActionExecutedEvent();
@@ -373,6 +472,11 @@ namespace Machina.Drivers.Communication
 
                 case ABBCommunicationProtocol.RES_EXTAX:
                     this.initExtAx = new ExternalAxes(data[0], data[1], data[2], data[3], data[4], data[5]);
+                    break;
+
+
+                case ABBCommunicationProtocol.RES_FULL_POSE:
+                    logger.Debug("received full pose: " + data.ToString());
                     break;
 
             }